Ts conversion for objdumpers
