+++ .eslintrc.yml
@@ -70,7 +70,7 @@ rules:
   no-useless-call: error
   no-useless-computed-key: error
   no-useless-concat: error
-  no-useless-constructor: error
+  '@typescript-eslint/no-useless-constructor': error
   no-useless-escape: error
   no-useless-rename: error
   no-useless-return: error
