+++ lib/objdumper/index.ts
